pyproject.toml
[project]
name = "ai-steward"
version = "2.0.0"
description = "Telegram-бот, обёртка над Claude Code CLI"
requires-python = ">=3.11"
dependencies = [
    "aiogram==3.15.0",
    "faster-whisper==1.1.0",
    "edge-tts==7.2.7",
    "pytesseract==0.3.13",
    "Pillow==11.1.0",
    "requests>=2.31.0",
    "pydantic-settings>=2.7.0",
    "python-json-logger==2.0.7",
    "aiosqlite==0.20.0",
    "sentry-sdk==2.19.2",
]

[project.scripts]
ai-steward = "ai_steward.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_steward"]
